refactor: move all logging functions to their own file
